beautiful generator version working
Next up is making copyable versions - that means no generators,
ever-ever! We can use yield from at the top level, but otherwise
everything needs to be done with explicit state.

The common pattern is going to be evaling each of the forms
that need to be evaled (eg: args to a function) and storing that
that evaluation is complete. Each __next__ function needs to
be able to figure out where we left off.